requirements.txt
oss2
cos-python-sdk-v5
pyyaml